refactor DiskInfo
