+++ master/src/test/java/org/apache/celeborn/service/deploy/master/clustermeta/ha/MasterStateMachineSuiteJ.java
@@ -40,7 +40,6 @@ import org.apache.celeborn.common.client.MasterClient;
 import org.apache.celeborn.common.identity.UserIdentifier;
 import org.apache.celeborn.common.meta.AppDiskUsageSnapShot;
 import org.apache.celeborn.common.meta.DiskInfo;
-import org.apache.celeborn.common.meta.DiskInfoBase;
 import org.apache.celeborn.common.meta.WorkerInfo;
 import org.apache.celeborn.common.quota.ResourceConsumption;
 import org.apache.celeborn.common.rpc.RpcEnv;
@@ -123,7 +122,7 @@ public class MasterStateMachineSuiteJ extends RatisBaseSuiteJ {
     HAMasterMetaManager masterStatusSystem = new HAMasterMetaManager(null, conf);
     File tmpFile = File.createTempFile("tef", "test" + System.currentTimeMillis());
 
-    Map<String, DiskInfoBase> disks1 = new HashMap<>();
+    Map<String, DiskInfo> disks1 = new HashMap<>();
     disks1.put("disk1", new DiskInfo("disk1", 64 * 1024 * 1024 * 1024L, 100, 100, 0));
     disks1.put("disk2", new DiskInfo("disk2", 64 * 1024 * 1024 * 1024L, 100, 100, 0));
     disks1.put("disk3", new DiskInfo("disk3", 64 * 1024 * 1024 * 1024L, 100, 100, 0));
@@ -148,7 +147,7 @@ public class MasterStateMachineSuiteJ extends RatisBaseSuiteJ {
             3,
             Collections.singletonMap("appId3", new ResourceConsumption(2000, 2, 2000, 2, null))));
 
-    Map<String, DiskInfoBase> disks2 = new HashMap<>();
+    Map<String, DiskInfo> disks2 = new HashMap<>();
     disks2.put("disk1", new DiskInfo("disk1", 64 * 1024 * 1024 * 1024L, 100, 100, 0));
     disks2.put("disk2", new DiskInfo("disk2", 64 * 1024 * 1024 * 1024L, 100, 100, 0));
     disks2.put("disk3", new DiskInfo("disk3", 64 * 1024 * 1024 * 1024L, 100, 100, 0));
@@ -173,7 +172,7 @@ public class MasterStateMachineSuiteJ extends RatisBaseSuiteJ {
             3,
             Collections.singletonMap("appId3", new ResourceConsumption(2000, 2, 2000, 2, null))));
 
-    Map<String, DiskInfoBase> disks3 = new HashMap<>();
+    Map<String, DiskInfo> disks3 = new HashMap<>();
     disks3.put("disk1", new DiskInfo("disk1", 64 * 1024 * 1024 * 1024L, 100, 100, 0));
     disks3.put("disk2", new DiskInfo("disk2", 64 * 1024 * 1024 * 1024L, 100, 100, 0));
     disks3.put("disk3", new DiskInfo("disk3", 64 * 1024 * 1024 * 1024L, 100, 100, 0));
@@ -411,7 +410,7 @@ public class MasterStateMachineSuiteJ extends RatisBaseSuiteJ {
     HAMasterMetaManager leaderStatusSystem = raftServers.get(0);
     HAMasterMetaManager followerStatusSystem1 = raftServers.get(1);
     HAMasterMetaManager followerStatusSystem2 = raftServers.get(2);
-    Map<String, DiskInfoBase> disks1 = new HashMap<>();
+    Map<String, DiskInfo> disks1 = new HashMap<>();
     Map<UserIdentifier, ResourceConsumption> userResourceConsumption1 = new HashMap<>();
 
     // per register produces 2 logs
